// src/app/interfaces/team-details.interface.ts
export interface ITeams {
    color?: string;
    fanBase?: string;
    fanClub?: string;
    headCoach?: string;
    highlights?: string;
    homeVenue?: string;
    icon?: string;
    profile?: string;
    rivals?: string;
    shortName?: string;
    teamId: string;
    teamName: string;
    twitterAccount?: string;
}